# backend/solarhub-nestjs/swagger-output.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/notification/subscribe": {
      "post": {
        "operationId": "DeviceController_subscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Notifications - Devices"
        ]
      }
    },
    "/api/v1/notification/unsubscribe": {
      "post": {
        "operationId": "DeviceController_unsubscribe",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications - Devices"
        ]
      }
    },
    "/api/v1/notification/devices": {
      "get": {
        "operationId": "DeviceController_getDevices",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications - Devices"
        ]
      }
    },
    "/api/v1/notification/tokens/{token_id}/deactivate": {
      "post": {
        "operationId": "DeviceController_deactivateDevice",
        "parameters": [
          {
            "name": "token_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Notifications - Devices"
        ]
      }
    },
...
    "/api/v1/companies/{company_id}/services": {
      "get": {
        "operationId": "CompanyServicesController_listServices",
        "parameters": [
          {
            "name": "company_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Services"
        ]
      },
      "post": {
        "operationId": "CompanyServicesController_createService",
        "parameters": [
          {
            "name": "company_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyServiceCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Services"
        ]
      }
    },
    "/api/v1/companies/{company_id}/services/{service_id}": {
      "put": {
        "operationId": "CompanyServicesController_updateService",
        "parameters": [
          {
            "name": "company_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyServiceUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Services"
        ]
      },
      "delete": {
        "operationId": "CompanyServicesController_deleteService",
        "parameters": [
          {
            "name": "company_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Services"
        ]
      }
    },
    "/api/v1/service-catalogs": {
      "get": {
        "operationId": "CompanyServiceCatalogsController_listCatalogs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Company Service Catalogs"
        ]
      }
    },
...
    "/api/v1/service-types": {
      "get": {
        "operationId": "ServiceTypesController_listServiceTypes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Service Types"
        ]
      },
      "post": {
        "operationId": "ServiceTypesController_createServiceType",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceTypeCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Service Types"
        ]
      }
    },
...
    "/api/v1/service-types/{id}": {
      "put": {
        "operationId": "ServiceTypesController_updateServiceType",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceTypeUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Service Types"
        ]
      },
      "delete": {
        "operationId": "ServiceTypesController_deleteServiceType",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Service Types"
        ]
      }
    },
...
    "/api/v1/offers/requests": {
      "post": {
        "operationId": "UserOfferRequestsController_createRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferRequestInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User Offer Requests"
        ]
      },
      "get": {
        "operationId": "UserOfferRequestsController_listMyRequests",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User Offer Requests"
        ]
      }
    },
...
    "/api/v1/offers/admin/offers": {
      "get": {
        "operationId": "AdminOffersController_adminListOffers",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Offers"
        ]
      }
    },
    "/api/v1/offers/admin/requests": {
      "get": {
        "operationId": "AdminOffersController_adminListRequests",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Offers"
        ]
      }
    },
...
  },
  "info": {
    "title": "SolarHub API",
    "description": "The SolarHub Backend API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "basic": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "SubscribeDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "platform": {
            "type": "string",
            "description": "'ios' or 'android'"
          },
          "device_id": {
            "type": "string"
          },
          "app_version": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "platform"
        ]
      },
...
      "CompanyServiceCreateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ]
      },
      "CompanyServiceUpdateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "ServiceTypeCreateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "ServiceTypeUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "OfferRequestInDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}